requirements.txt
evdev==1.6.1
paho-mqtt==1.6.1
pynput==1.7.6
python-xlib==0.33
six==1.16.0
